+++ libraries/WiFiClientSecure/src/ssl_client.cpp
@@ -202,7 +202,7 @@ int start_ssl_client(sslclient_context *ssl_client, const IPAddress& ip, uint32_
         }
         unsigned char psk[MBEDTLS_PSK_MAX_LEN];
         size_t psk_len = strlen(psKey)/2;
-        for (int j=0; j<strlen(psKey); j+= 2) {
+        for (size_t j=0; j<strlen(psKey); j+= 2) {
             char c = psKey[j];
             if (c >= '0' && c <= '9') c -= '0';
             else if (c >= 'A' && c <= 'F') c -= 'A' - 10;
